+++ src/components/orchestrator/Orchestrator.test.tsx
@@ -52,13 +52,16 @@ describe('Orchestrator', () => {
   const queryClient = new QueryClient()
   const OrchestratorTestWrapper = ({
     mode,
+    isDownloadEnabled = false,
   }: {
     mode: MODE_TYPE.COLLECT | MODE_TYPE.REVIEW | MODE_TYPE.VISUALIZE
+    isDownloadEnabled?: boolean
   }) => (
     <QueryClientProvider client={queryClient}>
       <Orchestrator
         metadata={metadata}
         mode={mode}
+        isDownloadEnabled={isDownloadEnabled}
         initialInterrogation={interrogation}
         // @ts-expect-error: we should have a better lunatic mock
         source={source}
@@ -258,5 +261,54 @@ describe('Orchestrator', () => {
 
     act(() => getByText('Continue').click())
     expect(document.title).not.toContain('*')
+  })
+
+  it('shows download button when enabled', async () => {
+    const user = userEvent.setup()
+    const { getByText } = renderWithRouter(
+      <OrchestratorTestWrapper
+        mode={MODE_TYPE.COLLECT}
+        isDownloadEnabled={true}
+      />,
+    )
+
+    act(() => getByText('Start').click())
+    act(() => getByText('Continue').click())
+
+    const e = getByText('my-question')
+
+    await user.click(e)
+    await user.keyboard('f')
+
+    act(() => getByText('Continue').click())
+
+    expect(getByText('Download data')).toBeInTheDocument()
+  })
+
+  it('always shows download button in visualize mode', async () => {
+    const user = userEvent.setup()
+    const { getByText } = renderWithRouter(
+      <OrchestratorTestWrapper mode={MODE_TYPE.VISUALIZE} />,
+    )
+
+    act(() => getByText('Start').click())
+    act(() => getByText('Continue').click())
+
+    const e = getByText('my-question')
+
+    await user.click(e)
+    await user.keyboard('f')
+
+    act(() => getByText('Continue').click())
+
+    expect(getByText('Download data')).toBeInTheDocument()
+  })
+
+  it('hides download button when disabled', async () => {
+    const { getByText, queryByText } = renderWithRouter(
+      <OrchestratorTestWrapper mode={MODE_TYPE.COLLECT} />,
+    )
+    act(() => getByText('Start').click())
+    expect(queryByText('Download data')).toBeNull()
   })
 })